Lab3 Checkpoint 1
This checkpoint should work in its entirety when connected to the jtag and OpalKelly Spartan6. It includes a simple counter. Additional modules needed for operation include generated ICON and ILA units, as well as a UCF (constraints) file with the clock pin connected. This should also give correct waveforms in ChipScope.